requirements.txt
Django==4.0
djangorestframework==3.13.1
gunicorn==20.1.0
mysqlclient==2.1.0
python-dotenv==0.19.2
django-extensions